backend/requirements.txt
flask==2.3.3
flask-cors==3.0.10
pymongo==4.4.0
python-dotenv==1.0.0
torch==2.2.0
transformers==4.41.2
pillow==10.0.1
pytesseract==0.3.10
bcrypt==4.0.1
PyJWT==2.8.0
uvicorn==0.23.2
